requirements.txt
fastapi
numpy
psycopg2-binary
sentence_transformers
uvicorn[standard]
google-genai
